Optimize generated code only when deploying an artifact
The generated code is recompiled to remove the debug symbols in order to
reduce the jar size. This invalidates the compileJava task, causing the
next build to do more work. This change restricts the optimization to be
run only when uploadArchives will be executed, which is used for snapshot
and release deployments.